Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-request-fallback.https-expected.txt

CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py? due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py? due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cannot load image https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE& due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cannot load image https://localhost:9443/service-workers/service-worker/resources/redirect.py?Redirect=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FPNGIMAGE%26 due to access control checks.

PASS initialize global state 
PASS The SW must intercept the request for a main resource. 
PASS The SW must intercept the request of same origin XHR. 
PASS The SW must intercept the request of CORS-unsupported other origin XHR. 
PASS The SW must intercept the request of CORS-supported other origin XHR. 
PASS The SW must intercept only the first request of redirected XHR. 
PASS The SW must intercept only the first request for XHR which is redirected to CORS-unsupported other origin. 
PASS The SW must intercept only the first request for XHR which is redirected to CORS-supported other origin. 
PASS The SW must intercept the request for image. 
PASS The SW must intercept the request for other origin image. 
PASS The SW must intercept the request for CORS-unsupported other origin image. 
PASS The SW must intercept the request for CORS-supported other origin image. 
PASS The SW must intercept only the first request for redirected image resource. 
PASS The SW must intercept only the first request for image resource which is redirected to other origin. 
PASS The SW must intercept only the first request for image resource which is redirected to CORS-unsupported other origin. 
PASS The SW must intercept only the first request for image resource which is redirected to CORS-supported other origin. 
PASS restore global state 
